+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './components/app';
+import App from './components/App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
